requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
python-multipart
firebase-admin==6.0.1
google-cloud-storage
google-cloud-firestore
google-cloud-vision
python-dotenv
requests
google-cloud-firestore
google-cloud-pubsub